refactor: extract shared barrels (helpers, root entry) + dedup dev-server
Continues the 66g condition-split consolidation, same surface-preserving shape
as the config barrel:

- helpers/index.shared.ts holds the ~43-line neutral re-export block; each entry
  keeps only its per-condition residue (handleServerAction binding; the
  client-only resolveComponentsClient, which must not be re-exported under
  react-server per bd-6pi).
- index.shared.ts factors the root '.' plugin entry into makeVitePluginReactServer
  (importContext); index.{server,client}.ts each bake in their side. Both still
  route through the neutral orchestrator dispatcher (no static-link hazard).
- dev-server/index.client.ts: drop a duplicate handleServerAction.client re-export.

Surface verified under both conditions: root '.' identical (4 names), ./helpers
40 shared + client-only resolveComponentsClient, ./dev-server unchanged name set.
build clean, 101 entrypoints, test:server 651/4-skip, test:unit 468.
