+++ plugin/dev-server/index.client.ts
@@ -1,6 +1,5 @@
 export * from "./handleServerAction.client.js";
 export * from "./configureReactServer.client.js";
 export * from "./cleanupServerAction.client.js";
-export * from "./handleServerAction.client.js";
 export * from "./configureRequestHandler.client.js";
 export * from "./plugin.client.js";
